% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{Bids}
\alias{Bids}
\title{BIDS Directory Handler}
\description{
Handles BIDS-formatted directories.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the BIDS root directory.}

\item{\code{raw_path}}{Path to the raw data directory.}

\item{\code{subjects}}{List containing subject objects found in the directory.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Bids-new}{\code{Bids$new()}}
\item \href{#method-Bids-print_tree}{\code{Bids$print_tree()}}
\item \href{#method-Bids-set_oblique}{\code{Bids$set_oblique()}}
\item \href{#method-Bids-get_slices}{\code{Bids$get_slices()}}
\item \href{#method-Bids-clone}{\code{Bids$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bids-new"></a>}}
\if{latex}{\out{\hypertarget{method-Bids-new}{}}}
\subsection{Method \code{new()}}{
Determine the path to the raw data folder.
It checks for a 'raw' subdirectory, otherwise uses the base path.


Load subject information from the raw data directory.
Identifies directories starting with 'sub-' and creates Subject objects.


Load derivative data for all subjects and sessions.
Looks for a 'derivatives' directory and processes each subject/session.


Initializes a BIDS object with the given path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bids$new(bids_root_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{bids_root_path}}{Path to the BIDS directory}

\item{\code{base_bids_path}}{Character string of the base BIDS directory path.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bids-print_tree"></a>}}
\if{latex}{\out{\hypertarget{method-Bids-print_tree}{}}}
\subsection{Method \code{print_tree()}}{
Prints a tree-like structure of the BIDS directory.
Shows subjects, sessions, and scans.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bids$print_tree(include_details = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{include_details}}{Logical, whether to include full file paths for
scans (default: FALSE).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bids-set_oblique"></a>}}
\if{latex}{\out{\hypertarget{method-Bids-set_oblique}{}}}
\subsection{Method \code{set_oblique()}}{
Set oblique center and affine transformation matrix from parcellation images matching a scan
pattern for each session.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bids$set_oblique(scan_pattern, folder_pattern = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scan_pattern}}{Regular expression pattern to match scan names}

\item{\code{folder_pattern}}{Optional regular expression pattern to match folder names}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of Slice objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bids-get_slices"></a>}}
\if{latex}{\out{\hypertarget{method-Bids-get_slices}{}}}
\subsection{Method \code{get_slices()}}{
Get standard slices from scans matching a scan pattern
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bids$get_slices(scan_pattern, folder_pattern = NULL, oblique_slices = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scan_pattern}}{Regular expression pattern to match scan names}

\item{\code{folder_pattern}}{Optional regular expression pattern to match folder names}

\item{\code{oblique_slices}}{Logical, whether to extract oblique slices using the oblique affine}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
List of Slice objects
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Bids-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Bids-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Bids$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
