% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice.R
\name{extract_oriented_slices}
\alias{extract_oriented_slices}
\title{Extract slices in canonical anatomical orientation using the affine matrix}
\usage{
extract_oriented_slices(img, rotation = NULL, center = NULL, spacing = 1)
}
\description{
Extract slices in canonical anatomical orientation using the affine matrix
}
