% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{Scan}
\alias{Scan}
\title{BIDS Scan Handler}
\description{
Represents a single scan file within a BIDS dataset.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Full path to the scan file (e.g., .nii or .nii.gz).}

\item{\code{scan_name}}{Descriptive name of the scan (e.g., "T1w").}

\item{\code{json_sidecar}}{List containing data from the JSON sidecar file, or NULL.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Scan-new}{\code{Scan$new()}}
\item \href{#method-Scan-format}{\code{Scan$format()}}
\item \href{#method-Scan-clone}{\code{Scan$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scan-new"></a>}}
\if{latex}{\out{\hypertarget{method-Scan-new}{}}}
\subsection{Method \code{new()}}{
Extracts the descriptive name of the scan from its filename.
Assumes BIDS naming convention (parts separated by '_').


Loads the JSON sidecar file associated with the scan.
Throws an error if the JSON file is invalid.


Initializes a scan object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scan$new(scan_file_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{scan_file_path}}{Character string path to the scan file.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Character string scan name (e.g., "T1w", "task-rest_bold").
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scan-format"></a>}}
\if{latex}{\out{\hypertarget{method-Scan-format}{}}}
\subsection{Method \code{format()}}{
Format the scan information for printing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scan$format()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Scan-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Scan-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Scan$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
