% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{Subject}
\alias{Subject}
\title{BIDS Subject Handler}
\description{
Represents a single subject within a BIDS dataset.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{path}}{Path to the subject's directory.}

\item{\code{id}}{Identifier for the subject (e.g., "01").}

\item{\code{sessions}}{List containing session objects for this subject.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Subject-new}{\code{Subject$new()}}
\item \href{#method-Subject-get_id}{\code{Subject$get_id()}}
\item \href{#method-Subject-get_name}{\code{Subject$get_name()}}
\item \href{#method-Subject-format}{\code{Subject$format()}}
\item \href{#method-Subject-clone}{\code{Subject$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subject-new"></a>}}
\if{latex}{\out{\hypertarget{method-Subject-new}{}}}
\subsection{Method \code{new()}}{
Load session information for the subject.
Identifies directories starting with 'ses-' within the subject's path.


Initializes a subject object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$new(subject_path)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject_path}}{Character string path to the subject's directory.}

\item{\code{subject_dir_path}}{Character string path to the subject's directory.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL. Populates the \code{self$sessions} list.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subject-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-Subject-get_id}{}}}
\subsection{Method \code{get_id()}}{
Get the subject identifier.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string subject ID (e.g., "01").
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subject-get_name"></a>}}
\if{latex}{\out{\hypertarget{method-Subject-get_name}{}}}
\subsection{Method \code{get_name()}}{
Get the full subject name (e.g., "sub-01").
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$get_name()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string subject name with prefix.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subject-format"></a>}}
\if{latex}{\out{\hypertarget{method-Subject-format}{}}}
\subsection{Method \code{format()}}{
Format the subject information for printing.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$format()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Character string representation (e.g., "Subject(id=01)").
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Subject-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Subject-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Subject$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
