% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bids.R
\name{validate_path}
\alias{validate_path}
\title{Validate path}
\usage{
validate_path(input_path)
}
\arguments{
\item{input_path}{Path to validate}
}
\value{
TRUE upon successful validation (otherwise throws error).
}
\description{
Classes and functions for handling BIDS-formatted directories

Verifies if a path exists as a file or directory.
}
\keyword{internal}
