% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/array_functions.r
\name{vector_principal_direction_with_sd}
\alias{vector_principal_direction_with_sd}
\title{Compute Principal Direction and Standard Deviation}
\usage{
vector_principal_direction_with_sd(x, na.rm = TRUE)
}
\arguments{
\item{x}{A 3D array representing vector data with dimensions (m, n, t), where t represents time.}

\item{na.rm}{A logical indicating whether NA values should be removed. Default is TRUE.}
}
\value{
A list containing two 3D arrays:
\item{principal_sd}{A 3D array of shape (m, n, 1), containing the standard deviation of principal directions.}
\item{orthogonal_sd}{A 3D array of shape (m, n, 2), containing the standard deviation of orthogonal directions.}
}
\description{
Computes the principal direction and standard deviation of a vector field using PCA across a 3D array.
}
